infra/darwin/api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
google-cloud-aiplatform==1.67.1
google-cloud-bigquery==3.25.0
google-cloud-secret-manager==2.21.1
vertexai==1.67.1
chromadb==0.4.15
openai==1.3.7
requests==2.31.0
pyyaml==6.0.1
python-multipart==0.0.6
numpy==1.24.3
